theme switch to component
