+++ app/components/layout/index.ts
@@ -3,5 +3,6 @@ export * from './content'
 export * from './error-boundary'
 export * from './footer'
 export * from './form'
-export * from './header'
+export * from './page-footer'
+export * from './page-header'
 export * from './side-nav'
