+++ app/components/templates/index.ts
@@ -9,4 +9,5 @@ export * from './spacer'
 export * from './status-button'
 export * from './textarea-field'
 export * from './toaster'
+export * from './theme-switch'
 export * from './user-dropdown'
